Moved external link definitions to navigation
